render html from tiptap json and organize tiptap files
