+++ confiture-web-app/src/components/report/ReportNotes.vue
@@ -3,8 +3,8 @@ import { computed } from "vue";
 
 import { useReportStore } from "../../store";
 import { isTiptapDocumentEmpty } from "../../utils";
+import TiptapRenderer from "../tiptap/TiptapRenderer.vue";
 import FileUpload from "../ui/FileUpload.vue";
-import Tiptap from "../ui/Tiptap.vue";
 
 const report = useReportStore();
 
@@ -13,11 +13,11 @@ const files = computed(() => report.data?.notesFiles);
 
 <template>
   <h2 class="fr-sr-only">Notes</h2>
-  <Tiptap
-    v-if="!isTiptapDocumentEmpty(report.data?.notes)"
-    :model-value="report.data?.notes"
-    :editable="false"
+  <TiptapRenderer
+    v-if="report.data?.notes && !isTiptapDocumentEmpty(report.data?.notes)"
+    :document="report.data?.notes"
   />
+
   <h3 v-if="files?.length" class="fr-text--sm">Pièces jointes</h3>
   <FileUpload
     v-if="files"
