+++ confiture-web-app/src/components/tiptap/TiptapButton.vue
